+++ src/controllers/ItemsController.php
@@ -155,6 +155,7 @@ class ItemsController extends BaseController
         $postItems = $request->getParam('items', [
             [
                 'elementId' => $request->getParam('elementId'),
+                'elementSiteId' => $request->getParam('elementSiteId'),
                 'fields' => $request->getParam('fields'),
                 'options' => $request->getParam('options'),
             ],
@@ -166,6 +167,7 @@ class ItemsController extends BaseController
 
         foreach ($postItems as $key => $postItem) {
             $elementId = $postItem['elementId'] ?? '';
+            $elementSiteId = $postItem['elementSiteId'] ?? '';
 
             if (!$elementId) {
                 $errors[$key] = new ItemError('Element ID must be provided.');
@@ -173,7 +175,7 @@ class ItemsController extends BaseController
                 continue;
             }
 
-            $element = Craft::$app->getElements()->getElementById($elementId);
+            $element = Craft::$app->getElements()->getElementById($elementId, null, $elementSiteId);
 
             if (!$element) {
                 $errors[$key] = new ItemError('Unable to find element.');
@@ -181,7 +183,7 @@ class ItemsController extends BaseController
                 continue;
             }
 
-            $item = $this->_setItemFromPost($elementId);
+            $item = $this->_setItemFromPost($elementId, $elementSiteId);
 
             // Check if we're allowed to manage lists
             $this->enforceEnabledList($item->getList());
@@ -262,6 +264,7 @@ class ItemsController extends BaseController
             [
                 'itemId' => $request->getParam('itemId'),
                 'elementId' => $request->getParam('elementId'),
+                'elementSiteId' => $request->getParam('elementSiteId'),
                 'options' => $request->getParam('options'),
                 'fields' => $request->getParam('fields'),
             ],
@@ -270,6 +273,7 @@ class ItemsController extends BaseController
         foreach ($postItems as $key => $postItem) {
             $itemId = $postItem['itemId'] ?? null;
             $elementId = $postItem['elementId'] ?? null;
+            $elementSiteId = $postItem['elementSiteId'] ?? null;
             $options = $postItem['options'] ?? [];
 
             if (!$elementId && !$itemId) {
@@ -291,6 +295,10 @@ class ItemsController extends BaseController
 
             if ($elementId) {
                 $query->elementId($elementId);
+            }
+
+            if ($elementSiteId) {
+                $query->elementSiteId($elementSiteId);
             }
 
             $item = $query->one();
@@ -352,6 +360,7 @@ class ItemsController extends BaseController
             [
                 'itemId' => $request->getParam('itemId'),
                 'elementId' => $request->getParam('elementId'),
+                'elementSiteId' => $request->getParam('elementSiteId'),
                 'options' => $request->getParam('options'),
                 'fields' => $request->getParam('fields'),
             ],
@@ -362,6 +371,7 @@ class ItemsController extends BaseController
         foreach ($postItems as $key => $postItem) {
             $itemId = $postItem['itemId'] ?? null;
             $elementId = $postItem['elementId'] ?? null;
+            $elementSiteId = $postItem['elementSiteId'] ?? null;
             $fields = $postItem['fields'] ?? [];
             $options = $postItem['options'] ?? [];
 
@@ -389,6 +399,10 @@ class ItemsController extends BaseController
                 $query->elementId($elementId);
             }
 
+            if ($elementSiteId) {
+                $query->elementSiteId($elementSiteId);
+            }
+
             $item = $query->one();
 
             if ($item) {
@@ -400,7 +414,7 @@ class ItemsController extends BaseController
 
                 $variables['items'][] = array_merge(['action' => 'removed'], $item->toArray());
             } else {
-                $item = $this->_setItemFromPost($elementId);
+                $item = $this->_setItemFromPost($elementId, $elementSiteId);
 
                 // Set any additional options and fields on the item
                 $item->setOptions($options);
@@ -429,6 +443,7 @@ class ItemsController extends BaseController
     {
         $request = Craft::$app->getRequest();
         $itemId = $request->getParam('itemId');
+        $options = $request->getParam('options');
 
         if (!$itemId) {
             return $this->returnError('Item ID must be provided.');
@@ -444,6 +459,8 @@ class ItemsController extends BaseController
         $this->enforceEnabledList($item->getList());
 
         $item->setFieldValuesFromRequest('fields');
+
+        $item->setOptions($options);
 
         if (!Wishlist::$plugin->getItems()->saveElement($item)) {
             return $this->returnError('Unable to update item in list.', ['item' => $item]);
@@ -492,10 +509,11 @@ class ItemsController extends BaseController
         $variables['fieldsHtml'] = $form->render();
     }
 
-    private function _setItemFromPost($elementId = null): Item
+    private function _setItemFromPost($elementId = null, $elementSiteId = null): Item
     {
         $request = Craft::$app->getRequest();
         $elementId = $request->getParam('elementId', $elementId);
+        $elementSiteId = $request->getParam('elementSiteId', $elementSiteId);
         $listId = $request->getParam('listId');
 
         $listTypeId = $request->getParam('listTypeId');
@@ -511,7 +529,7 @@ class ItemsController extends BaseController
         }
 
         // Create the item, and force the item's list to also be created if not already
-        return WishList::$plugin->getItems()->createItem($elementId, $listId, $listTypeId, true);
+        return WishList::$plugin->getItems()->createItem($elementId, $listId, $listTypeId, true, $elementSiteId);
     }
 
 }