+++ src/variables/WishlistVariable.php
@@ -41,7 +41,7 @@ class WishlistVariable
         return Item::find();
     }
 
-    public function item($elementId, $listId = null, $listType = null): ?Item
+    public function item($elementId, $listId = null, $listType = null, $elementSiteId = null): ?Item
     {
         $item = null;
         $listTypeId = null;
@@ -58,11 +58,11 @@ class WishlistVariable
         $list = Wishlist::$plugin->getLists()->getList($listId, false, $listTypeId);
 
         if ($list->id) {
-            $item = Item::find()->elementId($elementId)->listId($list->id)->listTypeId($listTypeId)->one();
+            $item = Item::find()->elementId($elementId)->elementSiteId($elementSiteId)->listId($list->id)->listTypeId($listTypeId)->one();
         }
 
         if (!$item) {
-            $item = WishList::$plugin->getItems()->createItem($elementId, $listId, $listTypeId);
+            $item = WishList::$plugin->getItems()->createItem($elementId, $listId, $listTypeId, false, $elementSiteId);
         }
 
         return $item;